feat: add alacritty-themes package to setup.sh
The `alacritty-themes` package has been added to the list of NPM packages to install in the `setup.sh` script. This package will be installed when running the script on an Arch-based system.

feat: add SSH agent and key setup to .zshrc

The `.zshrc` file has been updated to include the following changes:
- Start SSH agent in the background to manage SSH keys.
- Add SSH key to the agent for authentication.
- Define a function `start_tmux` to start a new tmux session under specific conditions.
- Call the `start_tmux` function to automatically start a tmux session when the conditions are met.